P1298 - MCU 0x0031 - Merged A/B Mode Fixes
Manually Merged in changes from the "Develop-for-John" branch that fixed
A/B mode operation.
Previously, the A/B mode select was made in the main loop.  It could
take up to 1.5ms to get to this point.  By this time, the lambda had
already reached EOC so if you were going from high to low energy you
would be stuck in high energy mode.
Move the A/B mode select to the trigger interrupt so that it always
happens directly after a pulse as intended.