P1298 - MCU 0x0032 -
Fixed Bug - The use of PIN = !PIN was causing a long process to change
the PIN value when compiled.  If an interrupt happened durring this
process and changes made to a pin on the port durring the interrupt
would be overwritten.  Change the C so that it compiled to code that
only affect a single PIN.

Added some debugging information
* Timing Error in INT1 Interrupt
* Number of fast restarts