P1298 - MCU 0x0033 Trying and stabilize system
Lots of little changes
1) Changed the WDT from 32ms to 4ms.  This is because some resets it
also requires a WDT mode.  This will make that reset faster.  Also, the
"fast reset" capacitor was discharing durring that 32ms.
Added more ClrWdt so that we don't have wdt resets durring operation
Also moved when the fast restart cap gets charged to earlier in the
lambda on process

2) Changed the LTC2656 DAC to not use the dac_clear line.  This becomes
an input on the pic and is wired high on the board

3) Changed the state that drives the Magnet Supply on off.  This
requires the dip switch on the magnet supply to be flipped as well

4) Faults - Removed all the magnet current faults for debugging and
running in local mode

4b) Faults - Removed the unused Lambda_VPeak under/over voltage from the
active faults