P1298 - A36760 Development Release V0xD145
Added the compile time option for to compile for A34760 or A36760 board.

Added the new pin mappings for the A36760 board.

NOTE!!! This is kinda hacky.  The code should work the same for both
boards, but any new features that touch the affected pins (including
this code) should be tested on both boards before being released.

NOTE.  LED2 has not been mapped yet because it will be a bit wonky and
require changes to C as well as pin definitions.