Development Release 0xD106
1) Change the Heater rollback curve to use the curve provided in the
attached data sheet.

2) Adjust how the multipliers are used so that one affects Portal/Gantry
mode and one affects Ultra Low dose

At the moment we use one of three multipliers (one is fixed value of 1.
The other two are adjustable).
In Portal/Gantry mode - High Energy - the calculated power is not
modified when reading the rolloff table.
In Portal/Gantry mode - Interleaved - the calcualted power is multiplied
by MULT1 before reading the rolloff table.
In Portal/Gantry mode - Low Energy - the calcualted power is multiplied
by MULT2 before reading the rolloff table.

we are going to keep two adjustable multipliers and use them as follows.
In Portal/Gantry mode - High Energy - the calculated power is not
modified when reading the rolloff table.
In Portal/Gantry mode - Interleaved - the calcualted power is multiplied
by MULT1 before reading the rolloff table.
In Portal/Gantry mode - Low Energy - This is not a "normal" mode of
operation.  The calculated power is NOT MODIFIED when reading the
rolloff table (we know this won't work, but don't care).

In Ultra Low Dose mode - High Energy - the calculated power is not
modified when reading the rolloff table.
In Ultra Low Dose mode - Interleaved - the calcualted power is
multiplied by MULT2 before reading the rolloff table.
In Ultra Low Dose mode - Low Energy - This is not a "normal" mode of
operation.  The calculated power is NOT MODIFIED when reading the
rolloff table (this probably won't work, but we don't care).

The Control board will arbitrate between Portal/Gantry mode and Ultra
Low Does mode based and High energy lambda set point.  If the High
energy lambda set point is less than 17500 Volts then we are in Ultra
Low Does mode